% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_a_level.R
\name{read_a_level}
\alias{read_a_level}
\title{Read HYDRUS-1D A_Level.out file}
\usage{
read_a_level(hydrus_output_path)
}
\arguments{
\item{hydrus_output_path}{Path to the directory with HYDRUS output.}
}
\value{
A data frame with columns:
  \describe{
    \item{time}{Simulation time [T]}
    \item{sum_r_top}{Cumulative potential surface flux at top boundary [L]. infiltration -/evaporation +}
    \item{sum_r_root}{Cumulative potential transpiration [L]}
    \item{sum_v_top}{Cumulative actual flux at surface [L]}
    \item{sum_v_root}{Cumulative actual transpiration [L]}
    \item{sum_v_bot}{Cumulative flux at bottom boundary [L]. inflow +/outflow -.}
    \item{h_top}{Pressure head at top node [L]}
    \item{h_root}{Mean pressure head in root zone [L]}
    \item{h_bot}{Pressure head at bottom node [L]}
    \item{a_level}{A-level index (integer; current variable boundary condition number)}
  }
}
\description{
Parses the A_Level.out output file produced by HYDRUS-1D and returns a
data frame with one row per time step.
}
