% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sas_input.R
\name{get_sas_input}
\alias{get_sas_input}
\title{Convert output from a HYDRUS model to input used in a SAS model}
\usage{
get_sas_input(hydrus_output_path, depths, times, node_spacing)
}
\arguments{
\item{hydrus_output_path}{path to directory where HYDRUS output is}

\item{depths}{numeric vector of the minimum and maximum soil depths of the HYDRUS model}

\item{times}{numeric vector of the minimum and maximum times to use for SAS input}

\item{node_spacing}{numeric vector of the}
}
\value{
a data frame with the following columns:
\describe{
   \item{t}{time}
   \item{S}{soil column storage}
   \item{J}{water input at the soil surface (precipitation & irrigation)}
   \item{ET}{evapotranspiration water output at the soil surface}
   \item{Q}{water output at the maximum depth at the soil column (deep percolation)}
}
}
\description{
Convert output from a HYDRUS model to input used in a SAS model
}
\examples{
get_sas_input("examples/hydrus_output",
                        depths = c(0, 150),
                        times = c(497, 861),
                        node_spacing = 1.75)
}
