% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sas_input.R
\name{get_sas_input}
\alias{get_sas_input}
\title{Convert output from a HYDRUS model to input used in a SAS model}
\usage{
get_sas_input(hydrus_output_path, depths, times, node_spacing)
}
\arguments{
\item{hydrus_output_path}{path to directory where HYDRUS output is}

\item{depths}{numeric vector of the minimum and maximum soil depths of the HYDRUS model}

\item{times}{numeric vector of the minimum and maximum times to use for SAS input}

\item{node_spacing}{numeric vector of the}
}
\description{
Convert output from a HYDRUS model to input used in a SAS model
}
