% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nod_inf.R
\name{read_nod_inf}
\alias{read_nod_inf}
\title{Read HYDRUS-1D Nod_Inf.out file}
\usage{
read_nod_inf(hydrus_output_path)
}
\arguments{
\item{hydrus_output_path}{Path to the directory with HYDRUS output}
}
\value{
A data frame with columns:
  \describe{
    \item{node}{Node index (integer)}
    \item{depth}{Node depth [L]}
    \item{head}{Pressure head [L]}
    \item{moisture}{Volumetric water content [-]}
    \item{k}{Hydraulic conductivity [L/T]}
    \item{c}{Specific water capacity [1/L]. The water age [T] is
             written instead when particle tracking is considered.}
    \item{flux}{Nodal value of Darcian velocity [L/T]}
    \item{sink}{Nodal value of root water uptake [1/T]}
    \item{kappa}{Kappa [-]}
    \item{v_ks_top}{v/KsTop ratio [-]}
    \item{temp}{Temperature [C]}
    \item{time}{Simulation time at which values were recorded [T]}
  }
}
\description{
Parses the Nod_Inf.out output file produced by HYDRUS-1D, which contains
nodal state variables at multiple saved time levels.  Returns a single
long-format data frame combining all time blocks.
}
